backend/requirements.txt
fastapi==0.104.1
uvicorn==0.20.0
uagents==0.12.0
scikit-learn==1.3.2
google-generativeai==0.3.2
requests==2.31.0
python-multipart==0.0.6
